+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Versions 1.1.2
 * Support for Fall of Oriath and the Harbinger league.
 * Added the new sounds for the PlayAlertSound rule.
+* The PlayAlertSound rule can now have "Orb" as its first value.
 * Removal of datasets as a feature with the end of the beta.
 
 ## Versions 1.1.0-1.1.1